requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
pusher==3.0.0
python-dotenv==0.19.0